Switch to eclipse-temurin:11-jre
https://www.infoq.com/news/2021/10/adoptium-releases-temurin-jdk/

Formerly known as AdoptOpenJDK, the independent organization changed
its name to Eclipse Adoptium after joining the Eclipse Foundation in
June 2020. The Adoptium Working Group was formally established nine
months later. The current AdoptOpenJDK website and API are still
available, although it’s advised to make use of the new Adoptium
website and API which is directly equivalent to the AdoptOpenJDK API
to enable a smooth transition.